+++ app/code/Magento/Sales/Model/Order/Pdf/Creditmemo.php
@@ -27,12 +27,12 @@ class Magento_Sales_Model_Order_Pdf_Creditmemo extends Magento_Sales_Model_Order
      * @param Magento_Payment_Helper_Data $paymentData
      * @param Magento_Core_Helper_Data $coreData
      * @param Magento_Core_Helper_String $coreString
-     * @param Magento_Core_Model_Store_Config $coreStoreConfig
-     * @param Magento_Core_Model_Config $coreConfig
+     * @param Magento_Core_Model_Store_ConfigInterface $coreStoreConfig
+     * @param Magento_Core_Model_Translate $translate
      * @param Magento_Core_Model_Dir $coreDir
      * @param Magento_Shipping_Model_Config $shippingConfig
-     * @param Magento_Core_Model_Translate $translate
-     * @param Magento_Sales_Model_Order_Pdf_TotalFactory $pdfTotalFactory
+     * @param Magento_Sales_Model_Order_Pdf_Config $pdfConfig
+     * @param Magento_Sales_Model_Order_Pdf_Total_Factory $pdfTotalFactory
      * @param Magento_Sales_Model_Order_Pdf_ItemsFactory $pdfItemsFactory
      * @param Magento_Core_Model_LocaleInterface $locale
      * @param Magento_Core_Model_StoreManagerInterface $storeManager
@@ -44,32 +44,32 @@ class Magento_Sales_Model_Order_Pdf_Creditmemo extends Magento_Sales_Model_Order
         Magento_Payment_Helper_Data $paymentData,
         Magento_Core_Helper_Data $coreData,
         Magento_Core_Helper_String $coreString,
-        Magento_Core_Model_Store_Config $coreStoreConfig,
-        Magento_Core_Model_Config $coreConfig,
+        Magento_Core_Model_Store_ConfigInterface $coreStoreConfig,
+        Magento_Core_Model_Translate $translate,
         Magento_Core_Model_Dir $coreDir,
         Magento_Shipping_Model_Config $shippingConfig,
-        Magento_Core_Model_Translate $translate,
-        Magento_Sales_Model_Order_Pdf_TotalFactory $pdfTotalFactory,
+        Magento_Sales_Model_Order_Pdf_Config $pdfConfig,
+        Magento_Sales_Model_Order_Pdf_Total_Factory $pdfTotalFactory,
         Magento_Sales_Model_Order_Pdf_ItemsFactory $pdfItemsFactory,
         Magento_Core_Model_LocaleInterface $locale,
         Magento_Core_Model_StoreManagerInterface $storeManager,
         array $data = array()
     ) {
+        $this->_locale = $locale;
+        $this->_storeManager = $storeManager;
         parent::__construct(
             $paymentData,
             $coreData,
             $coreString,
             $coreStoreConfig,
-            $coreConfig,
+            $translate,
             $coreDir,
             $shippingConfig,
-            $translate,
+            $pdfConfig,
             $pdfTotalFactory,
             $pdfItemsFactory,
             $data
         );
-        $this->_locale = $locale;
-        $this->_storeManager = $storeManager;
     }
 
     /**