+++ dev/tests/integration/testsuite/Magento/Newsletter/Model/QueueTest.php
@@ -57,7 +57,7 @@ class Magento_Newsletter_Model_QueueTest extends PHPUnit_Framework_TestCase
                 $objectManager->get('Magento_Core_Model_View_FileSystem'),
                 $objectManager->get('Magento_Core_Model_View_Design'),
                 $objectManager->get('Magento_Core_Model_Store_Config'),
-                $objectManager->get('Magento_Core_Model_Config')
+                $objectManager->get('Magento_Core_Model_Email_Template_Config'),
             )
         );
         $emailTemplate->expects($this->once())
@@ -107,7 +107,7 @@ class Magento_Newsletter_Model_QueueTest extends PHPUnit_Framework_TestCase
                 $objectManager->get('Magento_Core_Model_View_FileSystem'),
                 $objectManager->get('Magento_Core_Model_View_Design'),
                 $objectManager->get('Magento_Core_Model_Store_Config'),
-                $objectManager->get('Magento_Core_Model_Config')
+                $objectManager->get('Magento_Core_Model_Email_Template_Config'),
             )
         );
         $template->expects($this->any())->method('_getMail')->will($this->onConsecutiveCalls($mail, $brokenMail));